inference/requirements.txt
Pillow
numpy
tqdm
python-dotenv
pytoshop
google-genai
diffsynth
torch